issues resolved, committing final changes

+++ ecomm-api/handler/types.go
@@ -9,7 +9,7 @@ type ProductReq struct {
 	Description  string  `json:"description"`
 	Rating       int64   `json:"rating"`
 	NumReviews   int64   `json:"num_reviews"`
-	Price        float32 `json:"price"`
+	Price        float64 `json:"price"`
 	CountInStock int64   `json:"count_in_stock"`
 }
 
@@ -21,7 +21,7 @@ type ProductRes struct {
 	Description  string     `json:"description"`
 	Rating       int64      `json:"rating"`
 	NumReviews   int64      `json:"num_reviews"`
-	Price        float32    `json:"price"`
+	Price        float64    `json:"price"`
 	CountInStock int64      `json:"count_in_stock"`
 	CreatedAt    time.Time  `json:"created_at"`
 	UpdatedAt    *time.Time `json:"updated_at"`
@@ -30,16 +30,16 @@ type ProductRes struct {
 type OrderReq struct {
 	Items         []OrderItem `json:"items"`
 	PaymentMethod string      `json:"payment_method"`
-	TaxPrice      float32     `json:"tax_price"`
-	ShippingPrice float32     `json:"shipping_price"`
-	TotalPrice    float32     `json:"total_price"`
+	TaxPrice      float64     `json:"tax_price"`
+	ShippingPrice float64     `json:"shipping_price"`
+	TotalPrice    float64     `json:"total_price"`
 }
 
 type OrderItem struct {
 	Name      string  `json:"name"`
 	Quantity  int64   `json:"quantity"`
 	Image     string  `json:"image"`
-	Price     float32 `json:"price"`
+	Price     float64 `json:"price"`
 	ProductID int64   `json:"product_id"`
 }
 
@@ -47,9 +47,9 @@ type OrderRes struct {
 	ID            int64       `json:"id"`
 	Items         []OrderItem `json:"items"`
 	PaymentMethod string      `json:"payment_method"`
-	TaxPrice      float32     `json:"tax_price"`
-	ShippingPrice float32     `json:"shipping_price"`
-	TotalPrice    float32     `json:"total_price"`
+	TaxPrice      float64     `json:"tax_price"`
+	ShippingPrice float64     `json:"shipping_price"`
+	TotalPrice    float64     `json:"total_price"`
 	CreatedAt     time.Time   `json:"created_at"`
 	UpdatedAt     *time.Time  `json:"updated_at"`
 }
